requirements.txt
streamlit
python-dotenv
openai
google-generativeai
google-genai
requests